job #10183 update to output domain qualified names
When type references are output for interfaces in projects, they
sometimes need to be qualified by domain name. This change navigates
to the location of the type and finds the closest sibling component.
If the name of the type is not already qualified using '::' notation,
the name of the sibling component is prepended with '::'