job:  #12576  timer event nomenclature
This commit enables the translator to convert ASL event designations
into MASL event designations for event references passed to Set_Timer.
ASL passes 'event(<KL><event_number>)' to Set_Timer.  The parser was
not loading the markings on the events, so that was added.  The
mark information gives us what we need to find the proper event.