job:  #12576  type reassignment for out arguments
This commit recognizes that ASL positions all out arguments on the LHS
of an assignment statement.  These out arguments need to be typed based
upon the out parameters of the call expression.  This reassignment now
occurs in the statement_assign rendering.  This implies that local
declarations need to be resolved after statement rendering.  There
are some interesting exceptional cases for Get_Time_Remaining and
out arguments that are not locals but the out arguments of the
invoking service.