job:  #12576  enum literals in call expressions
Because enumerations are not fully resolvable by name alone, some types
will be mislinked.  The following will render to the correct types.
Get the type of the argument from the service signature.
Keep track of the position and use it to get the type positionally.
Render from the signature UDT + the literal enumerate item.