job #12728: update WASL output for collection types
ASL implicitly considers all structure typed parameters to be sets. ASL
does not support collection types other than with structures. In order
for MASL conversion to work, the parameter needs to be typed as a
collection. To support this, simply remove the 'sequence of' or 'set
of' from the front of a collection type on WASL output.